Fix pandoc finding logic when executed from bazel (#18)
`pypandoc` has a sophisticated logic to determine location of `pandoc` binary it depends on. When executed from `bazel` that logic does not work (none of the "guessed" locations work). This fix explicitly sets `PYPANDOC_PANDOC` env variable to point to the `pandoc` binary pulled by bazel, which is located in same directory as the root of this plugin (this is where it will be put by bazel during execution).

`PYPANDOC_PANDOC` is the only way for pypandoc to override any guessing logic and set explicitly the location of `pandoc` such that the resultant build is kept reproducible (i.e. it is guaranteed to use `pandoc` pulled by bazel and not any of the `pandoc` binaries which may have been preinstalled on the build machine).

The environment cannot be set outside of the python source file, because the plugin is started from `protoc` and it (protoc) defines the enfvironment the plugin binary will be executed in.